+++ duke.txt
@@ -1,4 +1,6 @@
-T | 0 |  123
-D | 0 |  submuit   |   1
-T | 0 |  1
+T | 1 |  12
 T | 0 |  12
+T | 0 |  1
+T | 1 |  1
+E | 0 |  task   | start to: end
+T | 0 |  123
