+++ production/CS2113_IP/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: Duke
+Main-Class: Sigma
