--- src/main/java/Deadline.java
@@ -1,7 +1,0 @@
-public class Deadline extends Task{
-    public Deadline(String description) {
-        super(description);
-        //match doBy
-    }
-
-}
