--- src/main/java/Event.java
@@ -1,6 +1,0 @@
-public class Event extends Task{
-    public Event(String description) {
-        super(description);
-        //match doBy
-    }
-}
