--- src/main/java/Todo.java
@@ -1,6 +1,0 @@
-public class Todo extends Task{
-    public Todo(String description) {
-        super(description);
-        //match doBy
-    }
-}
