+++ src/main/java/exceptions/DuplicateMarkException.java
@@ -1,2 +1,4 @@
+package exceptions;
+
 public class DuplicateMarkException extends Exception{
 }
