+++ src/main/java/exceptions/DuplicateUnmarkException.java
@@ -1,0 +1,4 @@
+package exceptions;
+
+public class DuplicateUnmarkException extends Exception{
+}
