+++ src/main/java/exceptions/EmptyTaskException.java
@@ -1,2 +1,4 @@
+package exceptions;
+
 public class EmptyTaskException extends Exception{
 }
