+++ src/main/java/tasks/Deadline.java
@@ -1,0 +1,9 @@
+package tasks;
+
+public class Deadline extends Task {
+    public Deadline(String description, boolean taskDone) {
+        super(description, taskDone);
+        //match doBy
+    }
+
+}
