+++ src/main/java/tasks/Event.java
@@ -1,0 +1,8 @@
+package tasks;
+
+public class Event extends Task {
+    public Event(String description, boolean taskDone) {
+        super(description, taskDone);
+        //match doBy
+    }
+}
