+++ src/main/java/tasks/Todo.java
@@ -1,0 +1,8 @@
+package tasks;
+
+public class Todo extends Task {
+    public Todo(String description, boolean taskDone) {
+        super(description, taskDone);
+        //match doBy
+    }
+}
